refactor/product search & cart. impl/usecart hook & useproductsearch hook
