+++ src/hooks/useCart.ts
@@ -1,0 +1,44 @@
+import { useDispatch, useSelector } from "react-redux";
+import { RootState, AppDispatch } from "@/redux/store";
+import { setCart } from "@/redux/slices/cartSlice";
+
+export function useCart() {
+  const dispatch = useDispatch<AppDispatch>();
+  const cart = useSelector((state: RootState) => state.cartState.cart);
+
+  const addToCart = (product: any) => {
+    const updatedCart = [...cart];
+    const existingItemIndex = updatedCart.findIndex(
+      (item) => item.id === product.id
+    );
+    if (existingItemIndex !== -1) {
+      updatedCart[existingItemIndex] = {
+        ...updatedCart[existingItemIndex],
+        quantity: updatedCart[existingItemIndex].quantity + 1,
+      };
+    } else {
+      updatedCart.push({ ...product, quantity: 1 });
+    }
+    dispatch(setCart(updatedCart));
+  };
+
+  const removeFromCart = (productId: number) => {
+    const updatedCart = cart.filter((item) => item.id !== productId);
+    dispatch(setCart(updatedCart));
+  };
+
+  const updateQuantity = (productId: number, newQuantity: number) => {
+    if (newQuantity < 1) return;
+    const updatedCart = cart.map((item) =>
+      item.id === productId ? { ...item, quantity: newQuantity } : item
+    );
+    dispatch(setCart(updatedCart));
+  };
+
+  return {
+    cart,
+    addToCart,
+    removeFromCart,
+    updateQuantity,
+  };
+}
