+++ src/redux/slices/cartSlice.ts
@@ -10,10 +10,12 @@ interface CartItem {
 
 interface CartState {
   cart: CartItem[];
+  createdAt: string | null;
 }
 
 const initialState: CartState = {
   cart: [],
+  createdAt: null,
 };
 
 const cartSlice = createSlice({
@@ -22,6 +24,12 @@ const cartSlice = createSlice({
   reducers: {
     setCart: (state, action: PayloadAction<CartItem[]>) => {
       state.cart = action.payload;
+      if (state.cart.length > 0 && !state.createdAt) {
+        state.createdAt = new Date().toLocaleString();
+      }
+      if (state.cart.length === 0) {
+        state.createdAt = null;
+      }
     },
     addToCart: (state, action: PayloadAction<CartItem>) => {
       const existingItemIndex = state.cart.findIndex(
@@ -32,9 +40,15 @@ const cartSlice = createSlice({
       } else {
         state.cart.push({ ...action.payload, quantity: 1 });
       }
+      if (state.cart.length === 1 && !state.createdAt) {
+        state.createdAt = new Date().toLocaleString();
+      }
     },
     removeFromCart: (state, action: PayloadAction<number>) => {
       state.cart = state.cart.filter((item) => item.id !== action.payload);
+      if (state.cart.length === 0) {
+        state.createdAt = null;
+      }
     },
     updateQuantity: (
       state,
